updates for 64-bit

+++ source/keys.cpp
@@ -41,13 +41,13 @@ auto key_state_t::down(key_t k) -> void
 auto key_state_t::on_key(key_t k, std::function<void()> const& fn) -> uint32
 {
 	down_events_.push_back(std::make_tuple(fooey::key_sequence_t(k), fn, 0u));
-	return down_events_.size();
+	return (uint32)down_events_.size();
 }
 
 auto key_state_t::on_key(key_combination_t const& k, std::function<void()> const& fn) -> uint32
 {
 	down_events_.push_back(std::make_tuple(fooey::key_sequence_t(k), fn, 0u));
-	return down_events_.size();
+	return (uint32)down_events_.size();
 }
 
 auto key_state_t::up(key_t k) -> void
